Refactor CI and OSDK tests, add OSDK publish workflow
