+++ .github/workflows/aster_ci.yml
@@ -1,4 +1,4 @@
-name: Integration test
+name: Asterinas CI
 
 on:
   pull_request:
@@ -7,7 +7,43 @@ on:
       - main
 
 jobs:
-  test:
+  aster-lint:
+    runs-on: ubuntu-latest
+    timeout-minutes: 10
+    container: asterinas/asterinas:0.3.0
+    steps:
+      - run: echo "Running in asterinas/asterinas:0.3.0"
+
+      - uses: actions/checkout@v3
+      
+      - run: make install_osdk
+
+      - name: Check
+        id: check
+        run: make check
+
+  aster-unit-test:
+    runs-on: ubuntu-latest
+    timeout-minutes: 15
+    container: asterinas/asterinas:0.3.0
+    steps:
+      - run: echo "Running in asterinas/asterinas:0.3.0"
+
+      - uses: actions/checkout@v3
+
+      - run: make install_osdk
+
+      - name: Usermode Unit test
+        id: usermode_unit_test
+        run: make test
+
+      - name: Ktest Unit Test
+        id: ktest_unit_test
+        run: make update_initramfs && make ktest
+
+      # TODO: add component check.
+
+  aster-integration-test:
     runs-on: ubuntu-latest
     timeout-minutes: 30
     container: asterinas/asterinas:0.3.1
@@ -49,4 +85,3 @@ jobs:
       - name: Regression Test (MicroVM)
         id: regression_test_linux
         run: make run AUTO_TEST=regression ENABLE_KVM=0 QEMU_MACHINE=microvm RELEASE_MODE=1
-
