+++ .github/workflows/docker_build.yml
@@ -6,7 +6,7 @@ on:
     paths:
       - VERSION
     branches:
-      - 'main'
+      - main
 
 jobs:
   docker:
