+++ osdk/.gitignore
@@ -8,6 +8,3 @@ target/
 
 # MSVC Windows builds of rustc generate these, which store debugging information
 *.pdb
-
-# Integration test generated files
-workspace_*
