+++ osdk/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cargo-osdk"
-version = "0.1.5"
+version = "0.2.0"
 edition = "2021"
 description = "Accelerate OS development with Asterinas OSDK"
 license = "MPL-2.0"
