+++ osdk/rust-toolchain.toml
@@ -1,3 +1,2 @@
 [toolchain]
-channel = "nightly-2023-12-01"
-components = ["rust-src", "rustc-dev", "llvm-tools-preview"]
+channel = "stable"
