+++ osdk/src/base_crate/Cargo.toml.template
@@ -1,0 +1,4 @@
+[package]
+name = "#NAME#"
+version = "#VERSION#"
+edition = "2021"
