+++ osdk/src/base_crate/mod.rs
@@ -4,9 +4,7 @@
 //! It will depend on the kernel crate.
 //!
 
-use std::{fs, path::Path, process, process::Command, str::FromStr};
-
-use crate::{error::Errno, error_msg};
+use std::{fs, path::Path, str::FromStr};
 
 pub fn new_base_crate(
     base_crate_path: impl AsRef<Path>,
@@ -17,17 +15,32 @@ pub fn new_base_crate(
         std::fs::remove_dir_all(&base_crate_path).unwrap();
     }
 
-    let mut cmd = Command::new("cargo");
-    cmd.arg("new").arg("--bin").arg(base_crate_path.as_ref());
-    cmd.arg("--vcs").arg("none");
+    let dep_crate_version = {
+        let cargo_toml = dep_crate_path.as_ref().join("Cargo.toml");
+        let cargo_toml = fs::read_to_string(cargo_toml).unwrap();
+        let cargo_toml: toml::Value = toml::from_str(&cargo_toml).unwrap();
+        let dep_version = cargo_toml
+            .get("package")
+            .unwrap()
+            .as_table()
+            .unwrap()
+            .get("version")
+            .unwrap()
+            .as_str()
+            .unwrap();
+        dep_version.to_string()
+    };
 
-    if !cmd.status().unwrap().success() {
-        error_msg!(
-            "Failed to create base crate at: {:#?}",
-            base_crate_path.as_ref()
-        );
-        process::exit(Errno::CreateBaseCrate as _);
-    }
+    // Create the directory
+    fs::create_dir_all(&base_crate_path).unwrap();
+    // Create the src directory
+    fs::create_dir_all(base_crate_path.as_ref().join("src")).unwrap();
+
+    // Write Cargo.toml
+    let cargo_toml = include_str!("Cargo.toml.template");
+    let cargo_toml = cargo_toml.replace("#NAME#", &(dep_crate_name.to_string() + "-osdk-bin"));
+    let cargo_toml = cargo_toml.replace("#VERSION#", &dep_crate_version);
+    fs::write(base_crate_path.as_ref().join("Cargo.toml"), cargo_toml).unwrap();
 
     // Set the current directory to the target osdk directory
     let original_dir = std::env::current_dir().unwrap();
@@ -44,7 +57,7 @@ pub fn new_base_crate(
     // Overrite the main.rs file
     let main_rs = include_str!("main.rs.template");
     // Replace all occurence of `#TARGET_NAME#` with the `dep_crate_name`
-    let main_rs = main_rs.replace("#TARGET_NAME#", &dep_crate_name.replace("-", "_"));
+    let main_rs = main_rs.replace("#TARGET_NAME#", &dep_crate_name.replace('-', "_"));
     fs::write("src/main.rs", main_rs).unwrap();
 
     // Add dependencies to the Cargo.toml
@@ -54,7 +67,7 @@ pub fn new_base_crate(
     copy_manifest_configurations(base_crate_path);
 
     // Get back to the original directory
-    std::env::set_current_dir(&original_dir).unwrap();
+    std::env::set_current_dir(original_dir).unwrap();
 }
 
 fn add_manifest_dependency(crate_name: &str, crate_path: impl AsRef<Path>) {
@@ -64,6 +77,14 @@ fn add_manifest_dependency(crate_name: &str, crate_path: impl AsRef<Path>) {
         let content = fs::read_to_string(mainfest_path).unwrap();
         toml::from_str(&content).unwrap()
     };
+
+    // Check if "dependencies" key exists, create it if it doesn't
+    if !manifest.contains_key("dependencies") {
+        manifest.insert(
+            "dependencies".to_string(),
+            toml::Value::Table(toml::Table::new()),
+        );
+    }
 
     let dependencies = manifest.get_mut("dependencies").unwrap();
