+++ osdk/src/commands/new/mod.rs
@@ -77,7 +77,7 @@ fn create_osdk_manifest(cargo_metadata: &serde_json::Value) {
     // The apt OVMF repo installs to `/usr/share/OVMF`
     fs::write(
         osdk_manifest_path,
-        r#"
+        r#"\
 [boot]
 ovmf = "/usr/share/OVMF"
 [qemu]
