--- osdk/src/test/mod.rs
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: MPL-2.0
-
-//! The crate unit test module
-
-mod cli;
-mod commands;
-mod config_manager;
-mod utils;
