+++ osdk/tests/bin.rs
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: MPL-2.0
+
+//! This module contains tests that invokes the `osdk` binary and checks the output.
+//! Please be sure the the `osdk` binary is built and available in the `target/debug`
+//! directory before running these tests.
+
+mod cli;
+mod commands;
+mod integration;
+mod util;
